add hooks for stake page
