+++ src/components-new-version/stake/stake-data/stake-form.tsx
@@ -22,7 +22,7 @@ export function StakeForm({ initialAmount = '' }: { initialAmount?: string }) {
   const [debouncedUpdate, setDebouncedUpdate] = useState<NodeJS.Timeout | null>(
     null
   )
-  const { isLoggedIn, sdkHasLoaded, walletAddress, setShowAuthFlow } =
+  const { isLoggedIn, sdkHasLoaded, walletAddress, primaryWallet, setShowAuthFlow } =
     useCurrentWallet()
 
   // Add token balance hooks for SSE token
@@ -32,7 +32,7 @@ export function StakeForm({ initialAmount = '' }: { initialAmount?: string }) {
     loading: inputBalanceLoading,
   } = useTokenBalance(walletAddress, SSE_MINT)
 
-  const { stake, isLoading: showStakeLoading } = useStake()
+  const { stake, isLoading: showStakeLoading } = useStake(primaryWallet, walletAddress)
 
   const validateAmount = (value: string): boolean => {
     if (value === '') return true
@@ -168,33 +168,12 @@ export function StakeForm({ initialAmount = '' }: { initialAmount?: string }) {
     }
   }
 
-  if (!sdkHasLoaded) {
-    return (
-      <div className="flex items-center justify-center gap-2">
-        <Spinner />
-        <p>{t('trade.checking_wallet_status')}</p>
-      </div>
-    )
-  }
-
-  if (!isLoggedIn) {
-    return (
-      <Button
-        variant={ButtonVariant.OUTLINE}
-        expand
-        onClick={() => setShowAuthFlow(true)}
-      >
-        {t('common.connect_wallet')}
-      </Button>
-    )
-  }
-
   return (
     <div>
       <div className="flex flex-col gap-2">
         <div className="flex justify-between items-center">
           <p>{t('common.amount')}</p>
-          {isLoggedIn && !inputBalanceLoading && inputBalance && (
+          {!inputBalanceLoading && inputBalance && (
             <div className="flex items-center gap-2">
               <p className="text-muted-foreground text-xs">
                 {t('common.balance')}: {inputBalance}
@@ -260,14 +239,38 @@ export function StakeForm({ initialAmount = '' }: { initialAmount?: string }) {
         {inputError && <p className="text-destructive">{inputError}</p>}
       </div>
 
-      <Button
-        expand
-        className="mt-4"
-        onClick={() => stake(displayAmount)}
-        disabled={showStakeLoading}
-      >
-        {showStakeLoading ? <Spinner /> : t('trade.stake')}
-      </Button>
-    </div>
+      {
+        !sdkHasLoaded ? (
+          <Button
+            variant={ButtonVariant.OUTLINE}
+            expand
+            className="mt-4"
+          >
+            <Spinner />
+            <p>{t('trade.checking_wallet_status')}</p>
+          </Button>
+        ) : (
+          !isLoggedIn ? (
+            <Button
+              variant={ButtonVariant.OUTLINE}
+              expand
+              className="mt-4"
+              onClick={() => setShowAuthFlow(true)}
+            >
+              {t('common.connect_wallet')}
+            </Button>
+          ) : (
+            <Button
+              expand
+              className="mt-4"
+              onClick={() => stake(displayAmount)}
+              disabled={showStakeLoading}
+            >
+              {showStakeLoading ? <Spinner /> : t('trade.stake')}
+            </Button>
+          )
+        )
+      }
+    </div >
   )
 }
